Fix color rendering for RGB+W Strips
Apparently, the RPI library interprets the alpha value of colors as the white component for RGB+W strips, so we need to adjust the values passed so that the white LEDs aren't always on at full brightness, as well as to prevent voltage drops because the strips are designed to have the white component being blended, versus having the RGB colors "pretened" to create white.

This fixes both of these issues.